Moved generateCombinedFiles task from BlitzPlugin to BlitzBasePlugin
